+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Merged [Playlist-Tools-SMP](https://github.com/regorxxx/Playlist-Tools-SMP) script.
 - Merged [Search-by-Distance-SMP](https://github.com/regorxxx/Search-by-Distance-SMP) script.
 - Merged [ListenBrainz-SMP](https://github.com/regorxxx/ListenBrainz-SMP) script.
+- Merged [Autobackup-SMP](https://github.com/regorxxx/Autobackup-SMP) script.
 ### Changed
 ### Removed
 ### Fixed